wat foutjes eruit gehaald

+++ src/Controllers/ModeratorController.cs
@@ -36,7 +36,7 @@ public class ModeratorController : Controller
 
         if (!String.IsNullOrEmpty(searchString))
         {
-            list = list.Where(c => c.IncomingClient.FirstName.ToLower().Substring(0, searchString.Length) == searchString.ToLower());
+            list = list.Where(c => c.IncomingClient.FirstName.ToLower().Contains(searchString.ToLower()));
         }
 
         if (list.Count() == 0)
@@ -92,12 +92,17 @@ public class ModeratorController : Controller
     public async Task<Boolean> BlockClient(string clientId)
     {
         var client = await _context.Clients.SingleOrDefaultAsync(c => c.Id == clientId);
+        if (client.IsBlocked) return false;
         var messages = await _context.Messages.Where(m => client.Id == m.Sender.Id && m.ReportCount > 0).ToListAsync();
 
         client.IsBlocked = true;
         _context.Messages.RemoveRange(messages);
         await _context.SaveChangesAsync();
-        // var input = await EmailPosting(client);
+
+        if (!EmailSender.ApiKeyIsNull())
+        {
+            var input = await EmailPosting(client);
+        }
 
         return true;
     }
